requirements.txt
dash==2.14.0
plotly==5.17.0
pandas==2.1.0
numpy==1.24.3
gunicorn==21.2.0
supabase==2.0.0
python-dotenv==1.0.0
